use private key if it exists for anonymous oodi

+++ scripts.sh
@@ -54,9 +54,14 @@ get_oodikone_server_backup() {
 }
 
 get_anon_oodikone() {
-    echo "$OODI_KEY" | awk  '{gsub("\\\\n","\n")};1' > private.key
-    chmod 400 private.key
-    
+    file=./private.key
+    if [ -e "$file" ]; then
+      echo "Private key exists"
+    else 
+      echo "No private key, echoing from environment variable OODI_KEY"
+      echo "$OODI_KEY" | awk  '{gsub("\\\\n","\n")};1' > private.key
+      chmod 400 private.key
+    fi
     GIT_SSH_COMMAND='ssh -i private.key' git clone git@github.com:UniversityOfHelsinkiCS/anonyymioodi.git
     mv anonyymioodi/anon.bak.bz2 ./$BACKUP_DIR/latest-pg.bak.bz2
     mv anonyymioodi/user-dump.bak.bz2 ./$BACKUP_DIR/latest-user-pg.bak.bz2
